requirements.txt
torch
probflow
gpytorch
gym
tqdm
matplotlib
gin-config
jupyter